Removed and added comments. Read description
Removed: warning comment at the start of the code (removes about 13
lines, calm down GDC)

Changed: Reassigned buttons 5 and 6 (LB/RB) to lowering/raising arms,
set buttons 7 and 8 (LT/RT) to the grab/release mech. (We totally
commented out the lower/raise to reassign 5 and 6 to the grab/release
mech to ensure the code worked and it was the Spike that was the
problem. Confirmed, we need a new Spike.)

Added: Comments to the grab/release mech code and an explanation of what
the gigantic commented block is. It's an exponential motor curve. We had
a few issues with it, if you want to fix and use it please do! Improved
the comment on the motor delay time (1ms update time!)