Changed README, and to explain clientside
As a result of some glitches and bugs while testing things, we were
forced to remove references to the extra class at the start of the
program. To save file size, we're just removing the whole class. The old
README promised a relay which we no longer have, so that was removed
from the README. I haven't documented the code well, so I don't remember
the PWM layout of everything. We are documenting and testing our code
and will publish our findings in an updated README.