Adding Camera Code
However, I will need some help with that.
The camera will be used to assist the driver with making sure the arms
are properly closed and will be displayed on the SmartDashboard.

The problem is I don't fully know how to implement it, and I am
currently reading the documentation to see if it will help.